Move internal component into a separate file
